# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ResolveAI - AI Customer Complaint Resolution Agent API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: complaints
    description: Complaint analysis and management
  - name: customers
    description: Customer profile operations
  - name: escalation
    description: Complaint escalation
  - name: dashboard
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyze:
    post:
      operationId: analyzeComplaint
      tags: [complaints]
      summary: Analyze a customer complaint
      description: Uses AI models to classify, assess sentiment, and generate resolution for a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeComplaintBody"
      responses:
        "200":
          description: Complaint analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeComplaintResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      description: Returns complaints processed, optionally filtered by company or customer
      parameters:
        - name: companyId
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnalyzeComplaintResponse"

  /complaints/{ticketId}:
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update a complaint (assign agent, change status, override resolution)
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintBody"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeComplaintResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"

  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  /companies/{companyId}/agents:
    get:
      operationId: listCompanyAgents
      tags: [companies]
      summary: List agents for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"

  /customer/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer profile
      description: Returns full customer profile with interaction history
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      description: Returns all customer profiles
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerSummary"

  /escalate:
    post:
      operationId: escalateComplaint
      tags: [escalation]
      summary: Escalate a complaint
      description: Manually escalate a complaint to a human agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateBody"
      responses:
        "200":
          description: Escalation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalateResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      description: Returns aggregated statistics for the dashboard
      parameters:
        - name: companyId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: string
      required:
        - error

    AnalyzeComplaintBody:
      type: object
      properties:
        complaint:
          type: string
          description: The complaint text from the customer
        customerId:
          type: string
          description: The customer ID
        companyId:
          type: string
          description: Optional target company id
        language:
          type: string
          description: BCP-47-like UI language code (en, hi, kn, ta, te, es, fr) used to localize the generated resolution text.
        imageBase64:
          type: string
          description: Optional base64-encoded image (data URL or raw base64) attached to the complaint for visual damage analysis.
        platform:
          type: string
          enum: [amazon, flipkart, swiggy, other]
          description: Optional source platform of the order.
        orderId:
          type: string
          description: Optional order/transaction ID supplied by the customer.
      required:
        - complaint
        - customerId

    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        specialty:
          type: string
      required: [id, name, specialty]

    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
        logo:
          type: string
        color:
          type: string
        slaHours:
          type: integer
        policies:
          type: array
          items:
            type: string
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
      required: [id, name, industry, logo, color, slaHours, policies, agents]

    UpdateComplaintBody:
      type: object
      properties:
        status:
          type: string
          enum: [resolved, pending, escalated, in_progress]
        assignedAgentId:
          type: ["string", "null"]
        resolutionOverride:
          type: ["string", "null"]

    ConfidenceBreakdown:
      type: object
      properties:
        classificationScore:
          type: number
        policyMatchScore:
          type: number
        historyFactor:
          type: number
        classificationWeight:
          type: number
        policyMatchWeight:
          type: number
        historyWeight:
          type: number
      required:
        - classificationScore
        - policyMatchScore
        - historyFactor
        - classificationWeight
        - policyMatchWeight
        - historyWeight

    EscalationDetail:
      type: object
      properties:
        assignedAgent:
          type: string
        reasons:
          type: array
          items:
            type: string
        summary:
          type: string
      required:
        - assignedAgent
        - reasons
        - summary

    AnalyzeComplaintResponse:
      type: object
      properties:
        ticketId:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        customerTier:
          type: string
        complaint:
          type: string
        complaintType:
          type: string
          enum: [billing, refund, technical, delivery, account, product_quality, unknown]
        severity:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        confidenceScore:
          type: number
        confidencePercentage:
          type: number
        confidenceLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        confidenceBreakdown:
          $ref: "#/components/schemas/ConfidenceBreakdown"
        sentiment:
          type: string
          enum: [positive, neutral, negative]
        frustrationScore:
          type: number
        resolution:
          type: string
        policyReference:
          type: string
        policyCode:
          type: string
        shouldEscalate:
          type: boolean
        escalation:
          oneOf:
            - $ref: "#/components/schemas/EscalationDetail"
            - type: "null"
        escalationSummary:
          type: ["string", "null"]
        assignedAgent:
          type: ["string", "null"]
        slaHours:
          type: integer
        createdAt:
          type: string
        companyId:
          type: string
        companyName:
          type: string
        status:
          type: string
          enum: [resolved, pending, escalated, in_progress]
        assignedAgentId:
          type: ["string", "null"]
        agentSpecialty:
          type: ["string", "null"]
        resolutionOverride:
          type: ["string", "null"]
        updatedAt:
          type: string
        authenticity:
          type: string
          enum: [genuine, suspicious, likely_fake]
        authenticityReasons:
          type: array
          items:
            type: string
        priorityRank:
          type: integer
        imageUrl:
          type: ["string", "null"]
          description: Stored image data URL for the complaint (if any).
        platform:
          type: ["string", "null"]
          enum: [amazon, flipkart, swiggy, other, null]
        orderId:
          type: ["string", "null"]
        imageAnalysis:
          oneOf:
            - $ref: "#/components/schemas/ImageAnalysisResult"
            - type: "null"
        fraudRisk:
          type: string
          enum: [low, medium, high]
        decision:
          type: string
          enum: [auto_refund, auto_resolve, escalate, request_more_info]
        decisionExplanation:
          type: string
      required:
        - ticketId
        - customerId
        - customerName
        - customerTier
        - complaint
        - complaintType
        - severity
        - confidenceScore
        - confidencePercentage
        - confidenceLevel
        - confidenceBreakdown
        - sentiment
        - frustrationScore
        - resolution
        - policyReference
        - policyCode
        - shouldEscalate
        - escalation
        - escalationSummary
        - assignedAgent
        - slaHours
        - createdAt

    ImageAnalysisResult:
      type: object
      properties:
        damageDetected:
          type: boolean
        damageType:
          type: string
          enum: [broken, scratched, wrong_item, stained, none]
        confidence:
          type: number
          description: 0-100
        fraudRisk:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
      required:
        - damageDetected
        - damageType
        - confidence
        - fraudRisk
        - notes

    HistoryItem:
      type: object
      properties:
        ticketId:
          type: string
        channel:
          type: string
          enum: [chat, email, phone]
        date:
          type: string
        issue:
          type: string
        complaintType:
          type: string
        status:
          type: string
          enum: [resolved, pending, escalated]
        resolution:
          type: string
        agentName:
          type: string
      required:
        - ticketId
        - channel
        - date
        - issue
        - complaintType
        - status
        - resolution
        - agentName

    Customer:
      type: object
      properties:
        customerId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        tier:
          type: string
          enum: [Premium, Standard, Basic]
        joinDate:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
      required:
        - customerId
        - name
        - email
        - phone
        - tier
        - joinDate
        - history

    CustomerSummary:
      type: object
      properties:
        customerId:
          type: string
        name:
          type: string
        email:
          type: string
        tier:
          type: string
          enum: [Premium, Standard, Basic]
        joinDate:
          type: string
        totalTickets:
          type: integer
        openTickets:
          type: integer
      required:
        - customerId
        - name
        - email
        - tier
        - joinDate
        - totalTickets
        - openTickets

    EscalateBody:
      type: object
      properties:
        ticketId:
          type: string
        customerId:
          type: string
        reason:
          type: string
      required:
        - ticketId
        - customerId
        - reason

    EscalateResponse:
      type: object
      properties:
        escalated:
          type: boolean
        assignedAgent:
          type: string
        summary:
          type: string
        eta:
          type: string
        ticketId:
          type: string
      required:
        - escalated
        - assignedAgent
        - summary
        - eta
        - ticketId

    SentimentTrendItem:
      type: object
      properties:
        hour:
          type: string
        positive:
          type: integer
        neutral:
          type: integer
        negative:
          type: integer
      required:
        - hour
        - positive
        - neutral
        - negative

    DashboardStats:
      type: object
      properties:
        totalToday:
          type: integer
        resolvedByAi:
          type: integer
        escalated:
          type: integer
        avgConfidence:
          type: number
        complaintsByType:
          type: object
          additionalProperties:
            type: integer
        sentimentTrend:
          type: array
          items:
            $ref: "#/components/schemas/SentimentTrendItem"
        resolvedRate:
          type: number
        avgFrustrationScore:
          type: number
      required:
        - totalToday
        - resolvedByAi
        - escalated
        - avgConfidence
        - complaintsByType
        - sentimentTrend
        - resolvedRate
        - avgFrustrationScore